// src/global/enums/wallet.enum.ts
export enum TOKEN_TYPES {
  ON_CHAIN = 'on-chain',
  OFF_CHAIN = 'off-chain',
}

export enum TOKEN_WITHDRAW_TYPES {
  ON_CHAIN = 'on-chain',
  OFF_CHAIN = 'off-chain',
}

export enum PLATFORMS {
  HOMNIFI = 'homnifi',
  HORYSMALL = 'horysmall',
  JB = 'jb',
  XPRO = 'xpro',
}

export enum CONVERSION_TYPES {
  CUSTOM = 'custom',
  DYNAMIC = 'dynamic',
}